Safari-611.1.10/LayoutTests/fast/dom/repeating-timer-mixed-visible-display-none-elements-throttling-expected.txt

Tests that a repeating timer changing the style of both visible and display:none elements does not throttled.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


The timer should initially not be throttled.
PASS internals.isTimerThrottled(timeoutId) is false
PASS internals.isTimerThrottled(timeoutId) is false
PASS successfullyParsed is true

TEST COMPLETE
Visible
